backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9

# AI & ML
anthropic==0.39.0
openai==1.54.0

# Database
supabase==2.9.1
psycopg2-binary==2.9.9

# Background Jobs
celery==5.4.0
redis==5.2.0

# File Processing
ffmpeg-python==0.2.0
python-magic==0.4.27

# Storage
boto3==1.35.0

# Authentication & Payments
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==11.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0